% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_pattern.R
\name{apply_pattern}
\alias{apply_pattern}
\title{Apply a pattern filter to a character vector.}
\usage{
apply_pattern(pattern, strings)
}
\arguments{
\item{pattern}{search_pattern.}

\item{strings}{character. The strings to filter down.}
}
\description{
Apply a pattern filter to a character vector.
}
