% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/director.R
\name{check_if_parser_and_preprocessor_are_identical}
\alias{check_if_parser_and_preprocessor_are_identical}
\title{If the parser and preprocessor for a path is the same, the user has probably made a mistake.}
\usage{
check_if_parser_and_preprocessor_are_identical(director, path)
}
\arguments{
\item{director}{director. The director object.}

\item{path}{character. The path to check for whether the preprocessor and
parser are identical.}
}
\value{
Nothing, but issue a warning in red crayon if they are identical,
  since it likely means the user forgot to specify a parser.
}
\description{
If the parser and preprocessor for a path is the same, the user has probably made a mistake.
}

