% Generated by roxygen2 (4.0.0): do not edit by hand
\name{compile}
\alias{compile}
\title{Compile a resource using a resource handler.}
\usage{
compile(resource_object, source_args, resource_name, tracking = TRUE)
}
\arguments{
\item{resource_object}{directorResource. The resource object to compile.}

\item{source_args}{list. The arguments to the \code{base::source} call.
Must include \code{local} of type \code{environment}.}

\item{resource_name}{character. The name of the resource. This
is used to fetch the resource handler.}

\item{tracking}{logical. Whether or not to perform modification tracking
by pushing accessed resources to the director's stack. The default is
\code{TRUE}.}
}
\value{
A list with keys \code{value, modified} containing the compiled
  value (some R object) and a \code{logical} flag that indicates whether
  or not any of the resources used have been modified.
}
\description{
Compile a resource using a resource handler.
}

