% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{complete_extension}
\alias{complete_extension}
\title{Complete the extension of a file (.r or .R).}
\usage{
complete_extension(name, base = NULL)
}
\arguments{
\item{name}{character. The filename sans extension.}

\item{base}{character. A base path to be prefixed to \code{name} when
checking if the suffixed versions exist. The final returned string will
not include this base.}
}
\value{
\code{name} suffixed by ".r" or ".R" according to which exists.
  (Many Unix-based systems are extension case-sensitive).
}
\description{
Complete the extension of a file (.r or .R).
}
\note{
This function assumes at least one file ending in .r or .R exists.
}
\examples{
\dontrun{
 # Assume we have a file \\code{"foo.R"}.
 stopifnot(complete_extension("foo") == "foo.R")

 # Assume we have a file \\code{"bar.r"}.
 stopifnot(complete_extension("bar") == "bar.R")
}
}

