\name{director-package}
\alias{director-package}
\alias{director}
\docType{package}
\title{
\packageTitle{director}
}
\description{
\packageDescription{director}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{director}
\packageIndices{director}

The main function is \code{\link{shootVideo}}.
}
\author{
\packageAuthor{director}

Maintainer: \packageMaintainer{director}
}
\keyword{ package }
