% Generated by roxygen2 (4.0.1): do not edit by hand
\name{director_exists}
\alias{director_exists}
\title{Determine whether a resource exists relative to a director object.}
\usage{
director_exists(resource, helper = FALSE)
}
\arguments{
\item{resource}{character. The name of the resource.}

\item{helper}{logical. Whether or not to check helper existence
in an idempotent resource. The default is \code{FALSE}.}
}
\value{
\code{TRUE} or \code{FALSE} according as it does or does not
  exist.
}
\description{
Determine whether a resource exists relative to a director object.
}
\examples{
\dontrun{
  # Imagine we have a file structure:
  #   - foo
  #     - one
  #       - one.R
  #       - helper.R
  #     - two.R
  #
  # Then the bellow will return \\code{TRUE}, \\code{FALSE}, and \\code{TRUE},
  # respectively. Note that the \\code{"helper.R"} file is not considered a
  # resource by the director as \\code{"one.R"} shares its name with its
  # parent directory and is considered the accessible resource.

  d <- director('foo')
  d$exists('one')
  d$exists('one/helper')
  d$exists('two')
}
}

