% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/director-filename.R
\name{director_filename}
\alias{director_filename}
\title{Convert a resource to a filename.}
\usage{
director_filename(name, absolute = FALSE, check.exists = TRUE,
  helper = FALSE, enclosing = FALSE)
}
\arguments{
\item{name}{character. The resource name to convert to a full path.}

\item{absolute}{character. Whether or not to return the absolute path
(i.e., prepended by the director root). The default is \code{FALSE}.}

\item{check.exists}{logical. Whether or not to check if the file exists.
The default is \code{TRUE}. This argument should be set to false if the
we are certain the file exists and can skip the check.}

\item{helper}{logical. Whether or not to handle helper files instead of
resources. The default is \code{FALSE}.}

\item{enclosing}{logical. If \code{TRUE} and \code{name} refers to an
idempotent resource the directory name will be returned instead of the
.R file.}
}
\value{
the absolute path, relative to the director root if
   \code{absolute = FALSE} and an absolute path if \code{absolute = TRUE}.
}
\description{
Convert a resource to a filename.
}

