% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{drop_idempotence}
\alias{drop_idempotence}
\title{Convert an idempotent resource name to a non-idempotent resource name.}
\usage{
drop_idempotence(filename)
}
\arguments{
\item{filename}{character. The filename to convert.}
}
\value{
the non-idempotent filename.
}
\description{
Convert an idempotent resource name to a non-idempotent resource name.
}

