% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{get_helpers}
\alias{get_helpers}
\title{Get all helper files associated with an idempotent resource directory.}
\usage{
get_helpers(path, ..., leave_idempotent = FALSE)
}
\arguments{
\item{path}{character. The *absolute* path of the idempotent resource.}

\item{...}{additional parameters to pass to \code{list.files}.}

\item{leave_idempotent}{logical. Whether or not to leave the
idempotent file (non-helper). By default \code{FALSE}.}
}
\value{
a character list of relative helper paths.
}
\description{
Get all helper files associated with an idempotent resource directory.
}
\examples{
\dontrun{
  # If we have a directory structure given by \\code{"model/model.R"},
  # \\code{"model/constants.R"}, \\code{"model/functions.R"}, then the
  # below will return \\code{c("constants.R", "functions.R")}.
  get_helpers("model")
}
}

