% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%<<\%}
\alias{\%<<\%}
\title{Append to a list or environment, overwriting if necessary.}
\usage{
obj1 \%<<\% obj2
}
\arguments{
\item{obj1.}{The object to be appended to.}

\item{obj2.}{The object to append.}
}
\description{
Append to a list or environment, overwriting if necessary.
}
\examples{
\dontrun{
  x <- list(a = 1)
  x \%<<\% list(b = 2) # list(a = 1, b = 2)
  x \%<<\% list(a = 2) # list(a = 2)
  y <- list2env(x)
  y \%<<\% list(b = 2) # environment with a = 1 and b = 2
  y \%<<\% list2env(list(b = 2)) # same as above
  y \%<<\% list(a = 2) # environment with a = 2
}
}
