% Generated by roxygen2 (4.0.1): do not edit by hand
\name{has_preprocessor}
\alias{has_preprocessor}
\title{Whether there exists a preprocessor for a resource.}
\usage{
has_preprocessor(resource_path)
}
\arguments{
\item{resource_path}{character. The resource name.}
}
\description{
Whether there exists a preprocessor for a resource.
}

