\name{initialize}
\alias{initialize}
\title{Initialize a director object.}
\usage{
initialize(root, project_name = root)
}
\arguments{
  \item{root}{character. The root directory for the
  director.}

  \item{project_name}{character. The name of the director
  project. Useful for error messages. For example, if a
  resource is not found, an error message to the effect of
  "no resource foo found in your \code{project_name}
  project" will be displayed.}
}
\description{
Initialize a director object.
}

