% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/zzy_director.R
\name{is.director}
\alias{is.director}
\title{Whether or not an object is a director.}
\usage{
is.director(x)
}
\arguments{
\item{x}{ANY.}
}
\description{
Whether or not an object is a director.
}

