% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/utils.r
\name{is.idempotent_directory}
\alias{is.idempotent_directory}
\title{Whether or not a directory is an idempotent resource.}
\usage{
is.idempotent_directory(dir)
}
\arguments{
\item{dir}{character. The directory to check.}
}
\value{
\code{TRUE} or \code{FALSE} according as the directory is idempotent.
  There is no checking to ensure the directory exists.
}
\description{
By definition, this means the directory contains a file with the same name
(ignoring extension) as the directory.
}
\examples{
\dontrun{
  # If we have a directory foo containing foo.R, then
  is.idempotent_directory('foo')
  # is TRUE, otherwise it's FALSE.
}
}

