\name{is.syberia_project}
\alias{is.syberia_project}
\title{Verifies that the given filename points to the base of a syberia project.}
\usage{
is.syberia_project(filename)
}
\arguments{
  \item{filename}{character. The directory supposedly
  containing a syberia project.}
}
\value{
TRUE or FALSE according as the directory is or is not a
Syberia project (including if the directory does not
exist).
}
\description{
Verifies that the given filename points to the base of a
syberia project.
}

