% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resource-preprocessor.R
\name{preprocessor}
\alias{preprocessor}
\alias{preprocessors}
\title{Apply the preprocessor to a resource.}
\usage{
preprocessor(object, ..., parse. = TRUE)
}
\arguments{
\item{object}{active_resource. See \code{\link{active_resource}}.}

\item{...}{additional parameters to pass to the next layer in the resource
parsing tower.}

\item{parse.}{logical. Whether or not to apply the \link{parser} to the
resource. If \code{FALSE}, this is equivalent to sourcing the resource's
file without running its parser. By default, \code{TRUE}.}
}
\value{
The parsed resource if \code{parse. = TRUE}, and the preprocessed
   resource otherwise.
}
\description{
Apply the preprocessor to a resource.
}
\note{
The parameters must be named \code{object} and \code{...} due to
   this method's inclusion in a \code{\link{tower}}.
}

