% Generated by roxygen2 (4.0.0): do not edit by hand
\name{register_parser}
\alias{register_parser}
\title{Register a resource parser.}
\usage{
register_parser(path, parser)
}
\arguments{
\item{path}{character. The prefix to look for in the director.}

\item{parser}{directorParser.}
}
\description{
Register a resource parser.
}

