% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/director-preprocessors.R
\name{register_preprocessor}
\alias{register_preprocessor}
\title{Register a resource preprocessor}
\usage{
register_preprocessor(path, preprocessor, overwrite = FALSE)
}
\arguments{
\item{path}{character. The prefix to look for in the director.}

\item{preprocessor}{function.}

\item{overwrite}{logical. If \code{TRUE}, \code{register_preprocessor} will overwrite
the route instead of erroring if the path already has a registered
preprocessor. The default is \code{FALSE}.}
}
\description{
Register a resource preprocessor
}
\examples{
\dontrun{
  d <- director("some/project")
  d$register_preprocessor('models', function() { print("I am a ", resource, ", a model!") })
  r <- d$resource("models/some_model.R")
  r$value() # Will print: I am a models/some_model, a model!
}
}

