% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utils.r
\name{resource_cache_key}
\alias{resource_cache_key}
\title{Create a resource cache key from a resource key.}
\usage{
resource_cache_key(resource_key)
}
\arguments{
\item{resource_key}{character. The resource key.}
}
\value{
a cache key, currently just \code{"resource_cache/"} followed by
   the \code{resource_key}.
}
\description{
This is the key under whose director cache the info about the resource
as of previous execution will be stored.
}

