% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resource-initialize.R
\name{resource_initialize}
\alias{resource_initialize}
\title{Initialize a director resource.}
\usage{
resource_initialize(director, name, provided_environment = new.env(parent =
  defining_environment), defining_environment = parent.frame())
}
\arguments{
\item{director}{director. The director object to which this resource
belongs.}

\item{name}{character. The name of the resource. For example, if
we have a standalone resource \code{"linear_model.R"}, then
its resource name would be \code{"linear_model"}. If we have
an idempotent resource \code{"ensemble_model"} (a directory),
its name would be \code{"ensemble_model"}.}

\item{provided_environment}{environment. An environment with values that
should always be provided to the resource upon sourcing. For example,
if you overwrote the formula operator \code{~} to provide some custom
syntax in your R script, you can put it in an environment and it will
be available when sourcing the resource. The default is a new
environment.}

\item{defining_environment}{environment. This will become the parent
environment of the \code{provided_environment}. If, for example,
the resource is created from a package method and uses functions
from within that package's Imports, the \code{defining_environment}
will ensure that compiling the resource will find those methods.}
}
\description{
Initialize a director resource.
}
\seealso{
\code{directorResource}
}

