% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{simple_cache}
\alias{simple_cache}
\title{A simple caching structure.}
\usage{
simple_cache()
}
\value{
A list of four methods \code{get}, \code{set}, \code{exists}
  and \code{unset} that modify another list under the hood.
}
\description{
A simple caching structure.
}

