% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strip_root}
\alias{strip_root}
\title{Strip a root file path from an absolute filename.}
\usage{
strip_root(root, filename)
}
\arguments{
\item{root}{character. The root path.}

\item{filename}{character. The full file name.}
}
\value{
the stripped file path.
}
\description{
Strip a root file path from an absolute filename.
}
\examples{
\dontrun{
  stopifnot("test" == strip_root("foo/bar/test", "test"))
}
}

