\docType{package}
\name{syberiaStructure}
\alias{syberiaStructure}
\alias{syberiaStructure-package}
\title{Syberia provides an opinionated unified framework for
fast iteration on classifier development and deployment. It is
founded on convention over configuration and aims to solve the
problems of classifier-specific data preparation and
classifier-specific modeling parameters.}
\description{
The syberiaStructure package is meant to offer a unified
set of tools for traversing common Syberia projects (on a
file-system level).
}

