% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.r
\name{try_memoize}
\alias{try_memoize}
\title{Attempt to memoize a function using the memoise package.}
\usage{
try_memoize(fn)
}
\arguments{
\item{fn}{function. The function to memoize.}
}
\value{
nothing, but \code{try_memoize} will use non-standard
  evaluation to memoize in the calling environment.
}
\description{
This function will load the \code{memoise} package if it is
available, or do nothing otherwise.
}
\details{
(Blame Hadley for the spelling of memoise.)
}

