% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/director-package.R, R/director.R
\docType{package}
\name{director}
\alias{director}
\alias{director-package}
\alias{director_}
\title{Management and Tracking of Files in R Projects.}
\format{\preformatted{List of 8
 $ new       :function (...)  
 $ classname : chr "director"
 $ public    :List of 30
  ..$ dependency_stack       : NULL
  ..$ cache                  : NULL
  ..$ initialize             :function (root, project_name = "")  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 16 15 52 1 15 1 16 52
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1237d13f8> 
  ..$ exists                 :function (resource, helper = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 27 20 68 1 20 1 27 68
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12287f958> 
  ..$ resource               :function (name, ..., defining_environment = parent.frame())  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 146 13 149 1 13 1 146 149
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x123748b50> 
  ..$ register_parser        :function (path, parser = function() \{
\}, overwrite = FALSE, cache = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 20 20 52 1 20 1 20 52
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1237a19c8> 
  ..$ register_preprocessor  :function (path, preprocessor, overwrite = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 15 26 39 1 26 1 15 39
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12377eed0> 
  ..$ has_preprocessor       :function (resource_path)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 46 21 48 1 21 1 46 48
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12377eed0> 
  ..$ find                   :function (pattern = "", method = "wildcard", base = "", by_mtime = TRUE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 56 18 88 1 18 1 56 88
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12282af40> 
  ..$ match_preprocessor     :function (resource_name)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 80 26 83 5 26 5 80 83
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ match_parser           :function (resource_name)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 84 20 87 5 20 5 84 87
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ preprocessor           :function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 88 20 88 63 20 63 88 88
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ parser                 :function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 89 14 89 51 14 51 89 89
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ cached_resources       :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 24 90 51 24 51 90 90
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ cache_get              :function (k)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 92 17 92 44 17 44 92 92
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ cache_set              :function (k, v)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 93 17 93 50 17 50 93 93
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ tracking_dependencies  :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 95 29 95 73 29 73 95 95
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ clear_resource_stack   :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 96 28 96 102 28 102 96 96
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ increment_nesting_level:function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 97 31 97 105 31 105 97 97
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ decrement_nesting_level:function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 98 31 98 105 31 105 98 98
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ nesting_level          :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 99 21 99 60 21 60 99 99
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ push_dependency        :function (dependency)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 100 23 102 5 23 5 100 102
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ peek_dependency_stack  :function (...)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 103 29 103 72 29 72 103 103
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ pop_dependency_stack   :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 104 28 104 64 28 64 104 104
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ empty_dependency_stack :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 105 30 105 68 30 68 105 105
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ root                   :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 107 20 107 39 20 39 107 107
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ project_name           :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 108 20 108 47 20 47 108 108
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ absolute               :function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 109 18 109 58 18 58 109 109
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ show                   :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 110 18 115 5 18 5 110 115
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x11eb28200> 
  ..$ filename               :function (name, absolute = FALSE, check.exists = TRUE, helper = FALSE, enclosing = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 16 22 43 1 22 1 16 43
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12285a190> 
 $ private   :List of 8
  ..$ .root                    : NULL
  ..$ .project_name            : NULL
  ..$ .resource_cache          : list()
  ..$ .registry                : NULL
  ..$ .dependency_nesting_level: num 0
  ..$ .parsers                 : list()
  ..$ .preprocessors           : list()
  ..$ .cached_resources        : list()
 $ active    : NULL
 $ inherit   : NULL
 $ parent_env:<environment: namespace:director> 
 $ lock      : logi TRUE
 - attr(*, "class")= chr "R6ClassGenerator"
}}
\usage{
director_
}
\description{
The director package is responsible for managing and loading resources in
some fixed directory structure. It has support for tracking changes between
consecutive loads of resources (so that we can tell if a script was modified
since we last ran it) and defining parsers that allow us to generalize from
the pernicious simple linear execution that is common to R.

A director is an \link[=https://github.com/wch/R6]{R6 class} responsible for
a collection of file-traversal related responsibilities.
}
\details{
Throughout, a "resource" refers to an R script
  with possible helper functions. A resource with helpers is identified
  by the following heuristic: if a filename sans extension is identical
  to its directory name, it is considered the primary accessible "resource"
  with any accompanying helpers (files in the same directory) natively
  invisible to the directory. For example, a file called \code{"foo.R"}
  residing in directory \code{"foo"} will be accessible from the director
  object using \code{director_object$resource('foo')}, but any other R scripts
  in the \code{"foo"} directory will not be visible to the director object.

  With this definition of resource, a director manages all of the following
  relative to a root directory.

\describe{
  \item{"Loading"}{Grabbing resources relative to the root directory
     using the \code{resource} method. This also provides information
     about the last time the resource was grabbed (modification time,
     previous body).}
  \item{"Tracking"}{Besides tracking information about the loaded resource
     (and any previous versions of the loaded resource), the director also
     maintains a stack of resources that have been loaded so far. This allows
     one to, for example, force clear the stack, execute some code, and have
     a list of resources that were relevant to the computation. The current
     stack is available with \code{director_object$stack(all = TRUE)}, which
     will also clear it.}
  \item{"Parsers"}{Some resources are not intended to be merely executed,
     but also parsed. For example, if we define several functions in an
     R script, we have access to those functions if we had sourced it
     with \code{base::source(script, local = some_environment)} as they
     are now present in \code{some_environment}. A parser is a wrapper
     around loading of resources that allows one to do some further
     computation with this information. For example, we may define
     a \code{read} and \code{write} function, but parse this information
     into some IO object when fetching the resource. Thus, a resource
     can function as inputs to some parser that produces some final
     resource object. Parsers can be defined for full directories or
     specific files.}
}
}
\keyword{datasets}

