% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/director-package.R, R/director.R, R/resource.R
\docType{package}
\name{director}
\alias{director}
\alias{director-package}
\alias{director_}
\title{Management and Tracking of Files in R Projects.}
\format{\preformatted{List of 8
 $ new       :function (...)  
 $ classname : chr "director"
 $ public    :List of 26
  ..$ dependency_stack       : NULL
  ..$ cache                  : NULL
  ..$ initialize             :function (root, project_name = "")  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 16 15 52 1 15 1 16 52
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f6f3ce0> 
  ..$ exists                 :function (resource, helper = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 27 20 68 1 20 1 27 68
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f73ee30> 
  ..$ resource               :function (name, provides = list(), body = TRUE, soft = FALSE, tracking = TRUE, helper = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 50 13 139 1 13 1 50 139
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10c15ca60> 
  ..$ virtual_resource       :function (name, defining_environment)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 80 21 97 1 21 1 80 97
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f463ca8> 
  ..$ register_parser        :function (path, parser = function() \{
\}, overwrite = FALSE, cache = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 20 20 51 1 20 1 20 51
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10c3f6f78> 
  ..$ register_preprocessor  :function (path, preprocessor, overwrite = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 15 26 38 1 26 1 15 38
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10c439578> 
  ..$ has_preprocessor       :function (resource_path)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 45 21 47 1 21 1 45 47
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10c439578> 
  ..$ find                   :function (pattern = "", method = "wildcard", base = "", by_mtime = TRUE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 56 18 88 1 18 1 56 88
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f7117f0> 
  ..$ match_preprocessor     :function (resource_name)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 80 26 82 5 26 5 80 82
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ match_parser           :function (resource_name)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 83 20 85 5 20 5 83 85
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ preprocessor           :function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 86 20 86 50 20 50 86 86
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ parser                 :function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 87 14 87 38 14 38 87 87
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ cached_resources       :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 88 24 88 51 24 51 88 88
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ tracking_dependencies  :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 90 29 90 73 29 73 90 90
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ clear_resource_stack   :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 91 28 91 102 28 102 91 91
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ increment_nesting_level:function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 92 31 92 105 31 105 92 92
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ decrement_nesting_level:function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 93 31 93 105 31 105 93 93
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ nesting_level          :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 94 21 94 60 21 60 94 94
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ push_dependency        :function (dependency)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 95 23 97 5 23 5 95 97
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ root                   :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 99 20 99 39 20 39 99 99
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ project_name           :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 100 20 100 47 20 47 100 100
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ absolute               :function (x)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 101 18 101 53 18 53 101 101
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ show                   :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 102 18 107 5 18 5 102 107
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f435bc8> 
  ..$ filename               :function (name, absolute = FALSE, check.exists = TRUE, helper = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 13 22 34 1 22 1 13 34
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f728278> 
 $ private   :List of 8
  ..$ .root                    : NULL
  ..$ .project_name            : NULL
  ..$ .resource_cache          : list()
  ..$ .registry                : NULL
  ..$ .dependency_nesting_level: num 0
  ..$ .parsers                 : list()
  ..$ .preprocessors           : list()
  ..$ .cached_resources        : list()
 $ active    : NULL
 $ inherit   : NULL
 $ parent_env:<environment: namespace:director> 
 $ lock      : logi TRUE
 - attr(*, "class")= chr "R6ClassGenerator"
}}
\usage{
director_
}
\description{
The director package is responsible for managing and loading resources in
some fixed directory structure. It has support for tracking changes between
consecutive loads of resources (so that we can tell if a script was modified
since we last ran it) and defining parsers that allow us to generalize from
the pernicious simple linear execution that is common to R.

A director is an \link[=https://github.com/wch/R6]{R6 class} responsible for
a collection of file-traversal related responsibilities.
}
\details{
Throughout, a "resource" refers to an R script
  with possible helper functions. A resource with helpers is identified
  by the following heuristic: if a filename sans extension is identical
  to its directory name, it is considered the primary accessible "resource"
  with any accompanying helpers (files in the same directory) natively
  invisible to the directory. For example, a file called \code{"foo.R"}
  residing in directory \code{"foo"} will be accessible from the director
  object using \code{director_object$resource('foo')}, but any other R scripts
  in the \code{"foo"} directory will not be visible to the director object.

  With this definition of resource, a director manages all of the following
  relative to a root directory.

\describe{
  \item{"Loading"}{Grabbing resources relative to the root directory
     using the \code{resource} method. This also provides information
     about the last time the resource was grabbed (modification time,
     previous body).}
  \item{"Tracking"}{Besides tracking information about the loaded resource
     (and any previous versions of the loaded resource), the director also
     maintains a stack of resources that have been loaded so far. This allows
     one to, for example, force clear the stack, execute some code, and have
     a list of resources that were relevant to the computation. The current
     stack is available with \code{director_object$stack(all = TRUE)}, which
     will also clear it.}
  \item{"Parsers"}{Some resources are not intended to be merely executed,
     but also parsed. For example, if we define several functions in an
     R script, we have access to those functions if we had sourced it
     with \code{base::source(script, local = some_environment)} as they
     are now present in \code{some_environment}. A parser is a wrapper
     around loading of resources that allows one to do some further
     computation with this information. For example, we may define
     a \code{read} and \code{write} function, but parse this information
     into some IO object when fetching the resource. Thus, a resource
     can function as inputs to some parser that produces some final
     resource object. Parsers can be defined for full directories or
     specific files.}
}
}
\keyword{datasets}

