% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/zzy_director.R
\name{director_.filename}
\alias{director_.filename}
\title{Convert a resource to a filename.}
\usage{
director_.filename(name, absolute = FALSE, check.exists = TRUE,
  helper = FALSE)
}
\arguments{
\item{name}{character. The resource name to convert to a full path.}

\item{absolute}{character. Whether or not to return the absolute path
(i.e., prepended by the director root). The default is \code{FALSE}.}

\item{check.exists}{logical. Whether or not to check if the file exists.
The default is \code{TRUE}. This should be primarily used if the file
has already been checked for existence.}

\item{helper}{logical. Whether or not to handle helper files instead of resources.
The default is \code{FALSE}.}
}
\value{
the full path, relative to the director root if \code{full = FALSE}
   and an absolute path if \code{FULL = TRUE}.
}
\description{
Convert a resource to a filename.
}

