% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_helpers}
\alias{get_helpers}
\title{Get all helper files associated with an idempotent resource directory.}
\usage{
get_helpers(path)
}
\arguments{
\item{path}{character. The *absolute* path of the idempotent resource.}
}
\value{
a character list of relative helper paths.
}
\description{
Get all helper files associated with an idempotent resource directory.
}
\examples{
\dontrun{
  # If we have a directory structure given by \\code{"model/model.R"},
  # \\code{"model/constants.R"}, \\code{"model/functions.R"}, then the
  # below will return \\code{c("constants.R", "functions.R")}.
  get_helpers("model")
}
}

