% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocessors.R
\name{has_preprocessor}
\alias{has_preprocessor}
\title{Whether there exists a preprocessor for a resource.}
\usage{
has_preprocessor(resource_path)
}
\arguments{
\item{resource_path}{character. The resource name.}
}
\value{
\code{TRUE} or \code{FALSE} depending on whether there
  is a preprocessor for this resource.
}
\description{
Whether there exists a preprocessor for a resource.
}

