% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/director-initialize.R
\name{initialize}
\alias{initialize}
\title{Initialize a director object.}
\usage{
initialize(root, project_name = "")
}
\arguments{
\item{root}{character. The root directory for the director.}

\item{project_name}{character. The name of the director project. Useful for
error messages. For example, if a resource is not found, an error message
to the effect of "no resource foo found in your \code{project_name}
project" will be displayed.}
}
\value{
a \code{director} reference class object.
}
\description{
Initialize a director object.
}
\examples{
\dontrun{
director(tempdir())
director(tempdir(), "my project") # Error messages on using the director's
                                  # methods will now usually result in
                                  # the ending "in project 'my project'".
}
}
