% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/director-parsers.R
\name{register_parser}
\alias{register_parser}
\title{Register a resource parser.}
\usage{
register_parser(path, parser = function() { }, overwrite = FALSE,
  cache = FALSE)
}
\arguments{
\item{path}{character. The prefix to look for in the director.}

\item{parser}{function.}

\item{overwrite}{logical. If \code{TRUE}, \code{register_parser} will overwrite
the route instead of erroring if the path already has a registered
parser. The default is \code{FALSE}.}

\item{cache}{logical. Whether or not to cache resources processed with this
parser. Cached resources will not be re-parsed if their dependencies have
not been modified. This distinction is important, as most resources are
factory resources (the object they generate should not be shared across
the entire project; instead, a copy should be made). The default is \code{FALSE}.}
}
\description{
Register a resource parser.
}
\examples{
\dontrun{
  d <- director('some/project')
  d$register_parser('models', function() { print("I am a ", resource, ", a model!") })
  r <- d$resource('models/some_model.R')
  r$value() # Will print: I am a models/some_model, a model!
}
}

