% Generated by roxygen2 (4.0.0): do not edit by hand
\name{register_parser}
\alias{register_parser}
\title{Register a resource parser.}
\usage{
register_parser(path, parser, overwrite = FALSE)
}
\arguments{
\item{path}{character. The prefix to look for in the director.}

\item{parser}{directorParser.}

\item{overwrrite}{logical. If \code{TRUE}, \code{register_parser} will overwrite
the route instead of erroring if the path already has a registered
parser. The default is \code{FALSE}.}
}
\description{
Register a resource parser.
}
\examples{
\dontrun{
  d <- director('some/project')
  d$register_parser('models', function() { print("I am a ", resource, ", a model!") })
  r <- d$resource('models/some_model.R')
  r$value() # Will print: I am a models/some_model, a model!
}
}

