% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{class}
\name{registry}
\alias{registry}
\alias{registry}
\title{A persistent on-disk cache of R objects associated with a directory.}
\description{
Having a registry attached to a project is very helpful for maintaining
state across R sessions without encoding everything in the unreliable
\code{.RData} file.
}
\details{
To create a registry, simply write \code{r <- registry("some/directory")}.
You can then use \code{r$set('some/key', some_value)} and
\code{r$get('some/key')} to set and retrieve values (arbitrary R objects).
}
\section{Methods}{

\describe{
\item{\code{initialize(root = NULL)}}{Initialize a registry.}
}}

\examples{
\dontrun{
  r <- registry('some/dir') # Create "some/dir" and make a registry there.
  r$set('some/key', value <- list(1,2,3))
  stopifnot(r$get('some/key'), value)
}
}
