% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utils.r
\name{resource_name}
\alias{resource_name}
\title{Convert a filename to a resource name.}
\usage{
resource_name(filename)
}
\arguments{
\item{filename}{character. The filename.}
}
\value{
the resource name (i.e., stripped of idempotence and extension).
}
\description{
Convert a filename to a resource name.
}

